requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asyncio==4.0.0
asyncpg==0.31.0
attrs==26.1.0
bcrypt==4.0.1
billiard==4.2.4
build==1.5.0
celery==5.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.9
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.7
dataclasses-json==0.6.7
distro==1.9.0
Django==6.0.3
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.129.0
filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.4.0
googleapis-common-protos==1.74.0
greenlet==3.3.1
groq==0.37.1
grpcio==1.80.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.14.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==7.1.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
kubernetes==35.0.0
langchain==1.2.17
langchain-classic==1.0.6
langchain-community==0.4.1
langchain-core==1.3.3
langchain-groq==1.1.2
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.25.1
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.9
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.3
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pypdf==6.10.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.27
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
regex==2026.4.4
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.52.1
stripe==15.1.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.8.0
typer==0.25.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unicorn==2.1.4
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
stripe